requirements.txt
# core libs
numpy>=1.24.3,<2.3.0
opencv-python==4.12.0.88
sounddevice==0.5.2
soundfile==0.13.1
pydub==0.25.1
simpleaudio==1.0.4
requests==2.31.0

# ML stack
torch==2.8.0
torchaudio==2.8.0

# Groq SDK
groq==0.31.0
python-dotenv==1.0.0
